[Refactor] Split runtime filter worker into focused modules
