+++ be/src/exec/pipeline/aggregate/aggregate_blocking_sink_operator.cpp
@@ -24,7 +24,7 @@
 #include "common/status.h"
 #include "exec/agg_runtime_filter_builder.h"
 #include "runtime/current_thread.h"
-#include "runtime/runtime_filter_worker.h"
+#include "runtime/runtime_filter_port.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks::pipeline {
