+++ be/src/exec/pipeline/hashjoin/hash_join_build_operator.cpp
@@ -25,7 +25,7 @@
 #include "exec/runtime_filter/runtime_filter_descriptor.h"
 #include "exec/runtime_filter/runtime_filter_probe.h"
 #include "runtime/current_thread.h"
-#include "runtime/runtime_filter_worker.h"
+#include "runtime/runtime_filter_port.h"
 #include "runtime/runtime_state.h"
 
 namespace starrocks::pipeline {
