+++ be/src/runtime/fragment_mgr.cpp
@@ -65,7 +65,7 @@
 #include "runtime/exec_env.h"
 #include "runtime/plan_fragment_executor.h"
 #include "runtime/runtime_filter_cache.h"
-#include "runtime/runtime_filter_worker.h"
+#include "runtime/runtime_filter_port.h"
 #include "runtime/runtime_metrics.h"
 #include "runtime/runtime_state_helper.h"
 #include "types/datetime_value.h"
