+++ be/src/runtime/runtime_filter_port.cpp
@@ -1,0 +1,254 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "runtime/runtime_filter_port.h"
+
+#include <algorithm>
+
+#include "common/config_network_fwd.h"
+#include "common/config_runtime_fwd.h"
+#include "exec/pipeline/hashjoin/hash_joiner_fwd.h"
+#include "exec/pipeline/runtime_filter_hub.h"
+#include "exec/runtime_filter/runtime_filter_descriptor.h"
+#include "exec/runtime_filter/runtime_filter_registry.h"
+#include "runtime/runtime_filter.h"
+#include "runtime/runtime_filter_cache.h"
+#include "runtime/runtime_filter_serde.h"
+#include "runtime/runtime_filter_worker.h"
+#include "runtime/runtime_filter_worker_event.h"
+#include "runtime/runtime_state.h"
+#include "runtime/service_contexts.h"
+#include "types/type_descriptor.h"
+
+namespace starrocks {
+
+namespace {
+
+const QueryExecutionServices& query_execution_services(RuntimeState* state) {
+    DCHECK(state != nullptr);
+    return *state->query_execution_services();
+}
+
+const RuntimeServices& runtime_services(RuntimeState* state) {
+    return *query_execution_services(state).runtime;
+}
+
+} // namespace
+
+std::string RuntimeFilterPort::listeners(int32_t filter_id) {
+    return _state->runtime_filter_registry()->waiters(filter_id);
+}
+
+void RuntimeFilterPort::publish_runtime_filters_for_skew_broadcast_join(
+        const std::list<RuntimeFilterBuildDescriptor*>& rf_descs_list, const std::vector<Columns>& key_columns,
+        const std::vector<bool>& null_safe, const std::vector<TypeDescriptor>& type_descs) {
+    pipeline::RuntimeMembershipFilters rf_descs(rf_descs_list.begin(), rf_descs_list.end());
+
+    for (const auto& rf_desc : rf_descs) {
+        auto* filter = rf_desc->runtime_filter();
+        if (filter == nullptr) continue;
+        _state->runtime_filter_port()->receive_runtime_filter(rf_desc->filter_id(), filter);
+    }
+
+    for (size_t i = 0; i < rf_descs.size(); i++) {
+        auto* rf_desc = rf_descs[i];
+        DCHECK(rf_desc->is_broad_cast_in_skew());
+        // Skip if runtime filter is nullptr or key columns are empty
+        if (rf_desc->runtime_filter() == nullptr || key_columns[i].empty()) {
+            continue;
+        }
+        // when enable_partitioned_hash_join is true, one runtime filter's key column can be split to multiple columns
+        // since skew broadcast join' build side data size is small, we just accumulate columns into a whole column for Convenience
+        auto column = key_columns[i][0]->clone();
+        for (size_t j = 1; j < key_columns[i].size(); j++) {
+            column->append(*key_columns[i][j]);
+        }
+        publish_skew_broadcast_join_key_columns(rf_desc, column, null_safe[i], type_descs[i]);
+    }
+}
+
+void RuntimeFilterPort::publish_runtime_filters(const std::list<RuntimeFilterBuildDescriptor*>& rf_descs) {
+    RuntimeState* state = _state;
+    for (auto* rf_desc : rf_descs) {
+        auto* filter = rf_desc->runtime_filter();
+        if (filter == nullptr) continue;
+        state->runtime_filter_port()->receive_runtime_filter(rf_desc->filter_id(), filter);
+    }
+    int timeout_ms = config::send_rpc_runtime_filter_timeout_ms;
+    if (state->query_options().__isset.runtime_filter_send_timeout_ms) {
+        timeout_ms = state->query_options().runtime_filter_send_timeout_ms;
+    }
+
+    int64_t rpc_http_min_size = config::send_runtime_filter_via_http_rpc_min_size;
+    if (state->query_options().__isset.runtime_filter_rpc_http_min_size) {
+        rpc_http_min_size = state->query_options().runtime_filter_rpc_http_min_size;
+    }
+
+    for (auto* rf_desc : rf_descs) {
+        auto* filter = rf_desc->runtime_filter();
+
+        if (filter == nullptr || !rf_desc->has_remote_targets()) continue;
+
+        // Empty runtime filter generated by broadcast join can not be used as a global runtime, because it
+        // maybe shirt-circuited by empty probe side.
+        if (filter->type() != RuntimeFilterSerializeType::IN_FILTER &&
+            rf_desc->join_mode() == TRuntimeFilterBuildJoinMode::BROADCAST &&
+            filter->get_membership_filter()->size() == 0)
+            continue;
+
+        auto directly_send_broadcast_grf = rf_desc->join_mode() == TRuntimeFilterBuildJoinMode::BROADCAST &&
+                                           !rf_desc->broadcast_grf_senders().empty();
+        // when send grf generated by broadcast join, GRF coordinator is needless.
+        if (!directly_send_broadcast_grf && rf_desc->merge_nodes().empty()) continue;
+
+        // for non-broadcast join, each fragment instance must send grf
+        // for broadcast join, if directly sending(not via GRF coordinator) is adopted, multiple fragment instances
+        // are chosen to send grf copies, otherwise, send only one copy. the senders are planned by FE.
+        auto need_sender_grf = rf_desc->join_mode() != TRuntimeFilterBuildJoinMode::BROADCAST ||
+                               rf_desc->broadcast_grf_senders().count(state->fragment_instance_id()) ||
+                               rf_desc->sender_finst_id() == state->fragment_instance_id();
+        if (!need_sender_grf) continue;
+
+        VLOG_FILE << "RuntimeFilterPort::publish_runtime_filters. join filter_id=" << rf_desc->filter_id()
+                  << ", finst_id=" << state->fragment_instance_id();
+
+        // rf metadata
+        PTransmitRuntimeFilterParams params;
+        params.set_transmit_timeout_ms(timeout_ms);
+        params.set_transmit_via_http_min_size(rpc_http_min_size);
+        prepare_params(params, state, rf_desc);
+
+        // print before setting data, otherwise it's too big.
+        VLOG_FILE << "RuntimeFilterPort::publish_runtime_filters. merge_node[0] = " << rf_desc->merge_nodes()[0]
+                  << ", query_id = " << params.query_id() << ", finst_id = " << params.finst_id()
+                  << ", be_number = " << params.build_be_number() << ", is_pipeline = " << params.is_pipeline()
+                  << ", filter = " << filter->debug_string();
+
+        std::string* rf_data = params.mutable_data();
+        size_t max_size = RuntimeFilterSerde::max_size(state, filter);
+        rf_data->resize(max_size);
+        size_t actual_size = RuntimeFilterSerde::serialize(state, filter, reinterpret_cast<uint8_t*>(rf_data->data()));
+        rf_data->resize(actual_size);
+
+        auto passthrough_delivery = actual_size <= config::deliver_broadcast_rf_passthrough_bytes_limit;
+        if (directly_send_broadcast_grf) {
+            auto sender_id =
+                    std::min_element(rf_desc->broadcast_grf_senders().begin(), rf_desc->broadcast_grf_senders().end(),
+                                     [](const auto& a, const auto& b) { return a.lo < b.lo; });
+            if (passthrough_delivery || *sender_id == state->fragment_instance_id()) {
+                runtime_services(state).runtime_filter_worker->send_broadcast_runtime_filter(
+                        std::move(params), rf_desc->broadcast_grf_destinations(), timeout_ms, rpc_http_min_size);
+            }
+        } else {
+            runtime_services(state).runtime_filter_worker->send_part_runtime_filter(
+                    std::move(params), rf_desc->merge_nodes(), timeout_ms, rpc_http_min_size, EventType::SEND_PART_RF);
+        }
+    }
+}
+
+void RuntimeFilterPort::publish_skew_broadcast_join_key_columns(RuntimeFilterBuildDescriptor* rf_desc,
+                                                                const ColumnPtr& keyColumn, bool null_safe,
+                                                                const TypeDescriptor& type_desc) {
+    DCHECK(rf_desc->join_mode() == TRuntimeFilterBuildJoinMode::BROADCAST);
+    DCHECK(!rf_desc->merge_nodes().empty());
+
+    RuntimeState* state = _state;
+    // only selected instance need to send rf
+    auto need_sender_grf = rf_desc->broadcast_grf_senders().count(state->fragment_instance_id());
+
+    if (!need_sender_grf) return;
+
+    int timeout_ms = config::send_rpc_runtime_filter_timeout_ms;
+    if (state->query_options().__isset.runtime_filter_send_timeout_ms) {
+        timeout_ms = state->query_options().runtime_filter_send_timeout_ms;
+    }
+
+    int64_t rpc_http_min_size = config::send_runtime_filter_via_http_rpc_min_size;
+    if (state->query_options().__isset.runtime_filter_rpc_http_min_size) {
+        rpc_http_min_size = state->query_options().runtime_filter_rpc_http_min_size;
+    }
+
+    PTransmitRuntimeFilterParams params;
+    params.set_transmit_timeout_ms(timeout_ms);
+    params.set_transmit_via_http_min_size(rpc_http_min_size);
+    prepare_params(params, state, rf_desc);
+
+    VLOG_FILE << "RuntimeFilterPort::publish_runtime_filters for skew join's broadcast site. join filter_id="
+              << rf_desc->filter_id() << ", finst_id=" << state->fragment_instance_id()
+              << "RuntimeFilterPort::publish_runtime_filters. merge_node[0]=" << rf_desc->merge_nodes()[0]
+              << ", query_id=" << params.query_id() << ", finst_id=" << params.finst_id()
+              << ", be_number=" << params.build_be_number() << ", is_pipeline=" << params.is_pipeline();
+
+    std::string* rf_data = params.mutable_data();
+    size_t max_size = RuntimeFilterSerde::skew_max_size(keyColumn);
+    rf_data->resize(max_size);
+    auto actual_size =
+            RuntimeFilterSerde::skew_serialize(keyColumn, null_safe, reinterpret_cast<uint8_t*>(rf_data->data()));
+    RETURN_IF(!actual_size.status().ok(), (void)nullptr);
+    rf_data->resize(actual_size.value());
+    *(params.mutable_columntype()) = type_desc.to_protobuf();
+    runtime_services(state).runtime_filter_worker->send_part_runtime_filter(std::move(params), rf_desc->merge_nodes(),
+                                                                            timeout_ms, rpc_http_min_size,
+                                                                            EventType::SEND_SKEW_JOIN_BROADCAST_RF);
+}
+
+void RuntimeFilterPort::prepare_params(PTransmitRuntimeFilterParams& params, RuntimeState* state,
+                                       RuntimeFilterBuildDescriptor* rf_desc) {
+    params.set_is_pipeline(rf_desc->is_pipeline());
+    params.set_filter_id(rf_desc->filter_id());
+    params.set_is_partial(true);
+    PUniqueId* query_id = params.mutable_query_id();
+    query_id->set_hi(state->query_id().hi);
+    query_id->set_lo(state->query_id().lo);
+    PUniqueId* finst_id = params.mutable_finst_id();
+    finst_id->set_hi(state->fragment_instance_id().hi);
+    finst_id->set_lo(state->fragment_instance_id().lo);
+    params.set_build_be_number(state->be_number());
+    params.set_is_skew_broadcast_join(rf_desc->is_broad_cast_in_skew());
+    if (rf_desc->is_broad_cast_in_skew()) {
+        params.set_skew_shuffle_filter_id(params.skew_shuffle_filter_id());
+    }
+}
+
+void RuntimeFilterPort::publish_local_colocate_filters(std::list<RuntimeFilterBuildDescriptor*>& rf_descs) {
+    RuntimeState* state = _state;
+    for (auto* rf_desc : rf_descs) {
+        auto* filter = rf_desc->runtime_filter();
+        if (filter == nullptr) continue;
+        state->runtime_filter_port()->receive_runtime_filter(rf_desc->filter_id(), filter);
+    }
+}
+
+void RuntimeFilterPort::receive_runtime_filter(int32_t filter_id, const RuntimeFilter* rf) {
+    runtime_services(_state).runtime_filter_cache->add_rf_event({
+            _state->query_id(),
+            filter_id,
+            "",
+            "LOCAL_PUBLISH",
+    });
+    VLOG_FILE << "RuntimeFilterPort::receive_runtime_filter(local). filter_id=" << filter_id
+              << ", waiters=" << _state->runtime_filter_registry()->waiters(filter_id)
+              << ", filter=" << rf->debug_string();
+    _state->runtime_filter_registry()->install_local(filter_id, rf);
+}
+
+void RuntimeFilterPort::receive_shared_runtime_filter(int32_t filter_id,
+                                                      const std::shared_ptr<const RuntimeFilter>& rf) {
+    VLOG_FILE << "RuntimeFilterPort::receive_runtime_filter(shared). filter_id=" << filter_id
+              << ", waiters=" << _state->runtime_filter_registry()->waiters(filter_id)
+              << ", filter=" << rf->debug_string();
+    _state->runtime_filter_registry()->install_shared(filter_id, rf);
+}
+
+} // namespace starrocks
