+++ be/src/runtime/runtime_filter_worker.h
@@ -14,211 +14,25 @@
 
 #pragma once
 
-#include <column/column.h>
-#include <types/logical_type.h>
-
-#include <atomic>
 #include <cstdint>
-#include <list>
-#include <map>
-#include <memory>
 #include <thread>
+#include <unordered_map>
 #include <vector>
 
-#include "base/brpc/ref_count_closure.h"
 #include "base/concurrency/blocking_queue.hpp"
-#include "base/uid_util.h"
-#include "common/object_pool.h"
-#include "common/status.h"
 #include "gen_cpp/InternalService_types.h"
 #include "gen_cpp/Types_types.h"
 #include "gen_cpp/internal_service.pb.h"
+#include "runtime/runtime_filter_delivery.h"
+#include "runtime/runtime_filter_merger.h"
+#include "runtime/runtime_filter_port.h"
 #include "runtime/runtime_filter_query_lifecycle.h"
-#include "runtime/runtime_filter_serde.h"
+#include "runtime/runtime_filter_worker_event.h"
+
 namespace starrocks {
-struct TypeDescriptor;
 
 struct RuntimeServices;
 struct RpcServices;
-class RuntimeState;
-
-class RuntimeFilter;
-class RuntimeFilterBuildDescriptor;
-
-using RuntimeFilterRpcClosure = RefCountClosure<PTransmitRuntimeFilterResult>;
-using RuntimeFilterRpcClosures = std::vector<RuntimeFilterRpcClosure*>;
-// RuntimeFilterPort is bind to a fragment instance
-// and it's to exchange RF(publish/receive) with outside world.
-class RuntimeFilterPort {
-public:
-    RuntimeFilterPort(RuntimeState* state) : _state(state) {}
-    void publish_runtime_filters(const std::list<RuntimeFilterBuildDescriptor*>& rf_descs);
-
-    void publish_runtime_filters_for_skew_broadcast_join(const std::list<RuntimeFilterBuildDescriptor*>& rf_descs,
-                                                         const std::vector<Columns>& keyColumns,
-                                                         const std::vector<bool>& null_safe,
-                                                         const std::vector<TypeDescriptor>& type_descs);
-
-    void publish_local_colocate_filters(std::list<RuntimeFilterBuildDescriptor*>& rf_descs);
-    // receiver runtime filter allocated in this fragment instance(broadcast join generate it)
-    // or allocated in this query(shuffle join generate global runtime filter)
-    void receive_runtime_filter(int32_t filter_id, const RuntimeFilter* rf);
-    void receive_shared_runtime_filter(int32_t filter_id, const std::shared_ptr<const RuntimeFilter>& rf);
-    std::string listeners(int32_t filter_id);
-
-private:
-    void publish_skew_broadcast_join_key_columns(RuntimeFilterBuildDescriptor* rf_desc, const ColumnPtr& keyColumn,
-                                                 bool null_safe, const TypeDescriptor& type_desc);
-    void static prepare_params(PTransmitRuntimeFilterParams& params, RuntimeState* state,
-                               RuntimeFilterBuildDescriptor* rf_desc);
-    RuntimeState* _state;
-};
-
-using SkewBroadcastRfMaterial = RuntimeFilterSkewMaterial;
-
-class RuntimeFilterMergerStatus {
-public:
-    RuntimeFilterMergerStatus() = default;
-    RuntimeFilterMergerStatus(RuntimeFilterMergerStatus&& other) noexcept
-            : arrives(std::move(other.arrives)),
-              expect_number(other.expect_number),
-              pool(std::move(other.pool)),
-              filters(std::move(other.filters)),
-              current_size(other.current_size),
-              max_size(other.max_size),
-              stop(other.stop),
-              is_skew_join(other.is_skew_join),
-              recv_first_filter_ts(other.recv_first_filter_ts),
-              recv_last_filter_ts(other.recv_last_filter_ts),
-              broadcast_filter_ts(other.broadcast_filter_ts) {}
-    // merge skew_broadcast_rf_material into out's _hash_partition_bf
-    Status _merge_skew_broadcast_runtime_filter(RuntimeFilter* out);
-    // which be number send this rf.
-    std::unordered_set<int32_t> arrives;
-    // how many partitioned rf we expect
-    int32_t expect_number;
-    ObjectPool pool;
-    // each partitioned rf.
-    std::map<int32_t, RuntimeFilter*> filters;
-    size_t current_size = 0;
-    size_t max_size = 0;
-    bool stop = false;
-    bool exceeded = true;
-    bool is_skew_join;
-
-    // only set when skew join optimization is used
-    SkewBroadcastRfMaterial* skew_broadcast_rf_material = nullptr;
-    // statistics.
-    // timestamp in ms since unix epoch;
-    // we care about diff not abs value.
-    int64_t recv_first_filter_ts = 0;
-    int64_t recv_last_filter_ts = 0;
-    int64_t broadcast_filter_ts = 0;
-
-    // only send once
-    bool isSent = false;
-};
-
-// RuntimeFilterMerger is to merge partitioned RF
-// and sent merged RF to consumer nodes.
-class RuntimeFilterMerger {
-public:
-    RuntimeFilterMerger(const RuntimeServices* runtime_services, const RpcServices* rpc_services,
-                        const UniqueId& query_id, const TQueryOptions& query_options, bool is_pipeline);
-    Status init(const TRuntimeFilterParams& params);
-    void merge_runtime_filter(PTransmitRuntimeFilterParams& params);
-    void store_skew_broadcast_join_runtime_filter(PTransmitRuntimeFilterParams& params);
-
-private:
-    void _send_total_runtime_filter(int rf_version, int32_t filter_id);
-    // filter_id -> where this filter should send to
-    std::map<int32_t, std::vector<TRuntimeFilterProberParams>> _targets;
-    std::map<int32_t, RuntimeFilterMergerStatus> _statuses;
-    const RuntimeServices* _runtime_services;
-    const RpcServices* _rpc_services;
-    UniqueId _query_id;
-    TQueryOptions _query_options;
-    const bool _is_pipeline;
-};
-
-enum EventType {
-    RECEIVE_TOTAL_RF = 0,
-    CLOSE_QUERY = 1,
-    OPEN_QUERY = 2,
-    RECEIVE_PART_RF = 3,
-    SEND_PART_RF = 4,
-    SEND_BROADCAST_GRF = 5,
-    SEND_SKEW_JOIN_BROADCAST_RF = 6,
-    RECEIVE_SKEW_JOIN_BROADCAST_RF = 7,
-    MAX_COUNT
-};
-
-inline std::string EventTypeToString(EventType type) {
-    switch (type) {
-    case RECEIVE_TOTAL_RF:
-        return "RECEIVE_TOTAL_RF";
-    case CLOSE_QUERY:
-        return "CLOSE_QUERY";
-    case OPEN_QUERY:
-        return "OPEN_QUERY";
-    case RECEIVE_PART_RF:
-        return "RECEIVE_PART_RF";
-    case SEND_SKEW_JOIN_BROADCAST_RF:
-        return "SEND_SKEW_JOIN_BROADCAST_RF";
-    case SEND_PART_RF:
-        return "SEND_PART_RF";
-    case SEND_BROADCAST_GRF:
-        return "SEND_BROADCAST_GRF";
-    case RECEIVE_SKEW_JOIN_BROADCAST_RF:
-        return "RECEIVE_SKEW_JOIN_BROADCAST_RF";
-    default:
-        break;
-    }
-    __builtin_unreachable();
-}
-// RuntimeFilterWorker works in a separated thread, and does following jobs:
-// 1. deserialize runtime filters.
-// 2. merge runtime filters.
-
-// it works in a event-driven way, and possible events are:
-// - create a runtime filter merger for a query
-// - receive partitioned RF, deserialize it and merge it, and sent total RF(for merge node)
-// - receive total RF and send it to RuntimeFilterPort
-// - send partitioned RF(for hash join node)
-// - close a query(delete runtime filter merger)
-struct RuntimeFilterWorkerEvent {
-    RuntimeFilterWorkerEvent() = default;
-    EventType type;
-    TUniqueId query_id;
-    // For OPEN_QUERY.
-    TQueryOptions query_options;
-    TRuntimeFilterParams create_rf_merger_request;
-    bool is_opened_by_pipeline;
-    // For SEND_PART_RF.
-    std::vector<TNetworkAddress> transmit_addrs;
-    std::vector<TRuntimeFilterDestination> destinations;
-    int transmit_timeout_ms;
-    int64_t transmit_via_http_min_size;
-    // For SEND_PART_RF, RECEIVE_PART_RF, and RECEIVE_TOTAL_RF.
-    PTransmitRuntimeFilterParams transmit_rf_request;
-};
-
-struct RuntimeFilterWorkerMetrics {
-    void update_event_nums(EventType event_type, int64_t delta) { event_nums[event_type] += delta; }
-
-    void update_rf_bytes(EventType event_type, int64_t delta) { runtime_filter_bytes[event_type] += delta; }
-
-    int64_t total_rf_bytes() {
-        int64_t total = 0;
-        for (int i = 0; i < EventType::MAX_COUNT; i++) {
-            total += runtime_filter_bytes[i];
-        }
-        return total;
-    }
-
-    std::array<std::atomic_int64_t, EventType::MAX_COUNT> event_nums{};
-    std::array<std::atomic_int64_t, EventType::MAX_COUNT> runtime_filter_bytes{};
-};
 
 class RuntimeFilterWorker : public RuntimeFilterQueryLifecycle {
 public:
@@ -242,30 +56,13 @@ public:
     const RuntimeFilterWorkerMetrics* metrics() const { return _metrics; }
 
 private:
-    void _receive_total_runtime_filter(PTransmitRuntimeFilterParams& params, int timeout_ms, int64_t rpc_http_min_size);
-    void _process_send_broadcast_runtime_filter_event(PTransmitRuntimeFilterParams&& params,
-                                                      std::vector<TRuntimeFilterDestination>&& destinations,
-                                                      int timeout_ms, int64_t rpc_http_min_size);
-    void _deliver_broadcast_runtime_filter_passthrough(PTransmitRuntimeFilterParams&& params,
-                                                       std::vector<TRuntimeFilterDestination>&& destinations,
-                                                       int timeout_ms, int64_t rpc_http_min_size);
-    void _deliver_broadcast_runtime_filter_relay(PTransmitRuntimeFilterParams&& params,
-                                                 std::vector<TRuntimeFilterDestination>&& destinations, int timeout_ms,
-                                                 int64_t rpc_http_min_size);
-    void _deliver_broadcast_runtime_filter_local(PTransmitRuntimeFilterParams& params,
-                                                 const TRuntimeFilterDestination& destinations, int timeout_ms,
-                                                 int64_t rpc_http_min_size);
-
-    void _deliver_part_runtime_filter(std::vector<TNetworkAddress>&& transmit_addrs,
-                                      PTransmitRuntimeFilterParams&& params, int transmit_timeout_ms,
-                                      int64_t rpc_http_min_size, const std::string& msg);
-
     bool _reach_queue_limit();
 
     UnboundedBlockingQueue<RuntimeFilterWorkerEvent> _queue;
     std::unordered_map<TUniqueId, RuntimeFilterMerger> _mergers;
     const RuntimeServices* _runtime_services;
     const RpcServices* _rpc_services;
+    RuntimeFilterDelivery _delivery;
     std::thread _thread;
     RuntimeFilterWorkerMetrics* _metrics = nullptr;
 };
